changes to controllers

+++ controllers/trainerController.js
@@ -5,7 +5,6 @@ const db = require("../models/index");
 
 router.get("/", (req, res) => {
   if (!req.session.currentUser) {
-    // if no currentuser cookie no access
     return res.redirect("/auth/login");
   }
   res.render("trainer/index");
@@ -13,23 +12,42 @@ router.get("/", (req, res) => {
 
 router.get("/new", (req, res) => {
   if (!req.session.currentUser) {
-    // if no currentuser cookie no access
     return res.redirect("/auth/login");
   }
   res.render("trainer/new");
 });
 
-router.post("/index", async (req, res) => {
+router.post("/", async (req, res) => {
   if (!req.session.currentUser) {
-    // if no currentuser cookie no access
     return res.redirect("/auth/login");
   }
-  const newTrainer = db.Trainer.create(req.body);
+  try {
+    const trainerData = {
+      trainerName: req.body.trainerName,
+      favPokemon: req.body.favPokemon,
+      goTeam: req.body.goTeam,
+      user: req.session.currentUser,
+      aboutMe: req.body.aboutMe,
+    };
+    const newTrainer = await db.Trainer.create(trainerData);
+    const foundUser = await db.User.findById(req.session.currentUser);
+    console.log(newTrainer);
+    console.log(foundUser);
+    await foundUser.trainers.push(newTrainer._id);
+    foundUser.save();
+
+    res.redirect("/pokemon");
+  } catch (err) {
+    return res.send(err);
+  }
 });
 
 router.get("/:id", async (req, res) => {
   try {
     const foundTrainer = await db.Trainer.findById(req.params.id);
+    if (req.session.currentUser !== foundTrainer.user) {
+      return res.redirect("/auth/login");
+    }
     res.render("trainer/show", {
       trainer: foundTrainer,
     });
@@ -41,6 +59,9 @@ router.get("/:id", async (req, res) => {
 router.get("/:id/edit", async (req, res) => {
   try {
     const foundTrainer = await db.Trainer.findById(req.params.id);
+    if (req.session.currentUser !== foundTrainer.user) {
+      return res.redirect("/auth/login");
+    }
     res.render("trainer/edits", {
       trainer: foundTrainer,
     });
@@ -49,7 +70,25 @@ router.get("/:id/edit", async (req, res) => {
   }
 });
 
+router.get("/:id/add", async (req, res) => {
+  try {
+    const foundTrainer = await db.Trainer.findById(req.params.id);
+    // if (req.session.currentUser !== foundTrainer.user) {
+    //   return res.redirect("/auth/login");
+    // }
+    res.render("trainer/add", {
+      trainer: foundTrainer,
+    });
+  } catch (err) {
+    return res.send(err);
+  }
+});
+
 router.put("/:id", async (req, res) => {
+  // if (!req.session.currentUser) {
+  //   return res.redirect("/auth/login");
+  // }
+  console.log(req.body)
   try {
     const foundTrainer = await db.Trainer.findByIdAndUpdate(
       req.params.id,
@@ -63,8 +102,14 @@ router.put("/:id", async (req, res) => {
 });
 
 router.delete("/:id", async (req, res) => {
+  if (!req.session.currentUser) {
+    return res.redirect("/auth/login");
+  }
   try {
-    await db.Trainer.findByIdAndDelete(req.params.id);
+    const deletedTrainer = await db.Trainer.findByIdAndDelete(req.params.id);
+    const foundUser = await db.User.findById(deletedTrainer.user);
+    foundUser.trainers.remove(deletedTrainer);
+    foundUser.save();
   } catch (error) {
     return res.send(err);
   }
